ADD split KV cache for Attention kernel (#579)
* init

* fix

* Integrate to MPK demo, illegal memory access to be fixed (#29)

* fix

* Update split kv attention (#30)

* Integrate to MPK demo, illegal memory access to be fixed

* checkpoint

* fix

* batch size 8 works

* add batch 1

* Splitkv demo finished (#31)

* Integrate to MPK demo, illegal memory access to be fixed

* checkpoint

* fix

* batch size 8 works

* comment out unused code

* format

* opt merge kernel

* fix union

* Add attention split kv hopper (#32)

* Add split kv attention for hopper

* aligned

* change base in attention hopper

* change request_id to int16_t and move it into union task_metadata

* format

---------

+++ tests/runtime_python/runtime_kernel_wrapper.cu
@@ -506,7 +506,7 @@ __global__ void multitoken_paged_attention_wrapper(
     int const *paged_kv_indptr_buffer_ptr,
     int const *paged_kv_indices_buffer_ptr,
     int const *paged_kv_last_page_len_buffer_ptr,
-    int request_id,
+    int16_t request_id,
     bool qk_norm,
     bool rope,
     void const *q_norm_weight_ptr,
@@ -563,7 +563,7 @@ void launch_multitoken_paged_attention(
     int const *paged_kv_indptr_buffer_ptr,
     int const *paged_kv_indices_buffer_ptr,
     int const *paged_kv_last_page_len_buffer_ptr,
-    int request_id,
+    int16_t request_id,
     bool qk_norm,
     bool rope,
     void const *q_norm_weight_ptr,
@@ -728,7 +728,7 @@ void multitoken_paged_attention(
     torch::Tensor paged_kv_indptr_buffer,
     torch::Tensor paged_kv_indices_buffer,
     torch::Tensor paged_kv_last_page_len_buffer,
-    int request_id,
+    int16_t request_id,
     bool qk_norm,
     bool rope,
     torch::optional<torch::Tensor> q_norm_weight = torch::nullopt,
@@ -1666,10 +1666,7 @@ __global__ void sampling_from_logits_test_wrapper(DType *logits,
                                                   uint64_t philox_seed,
                                                   uint64_t philox_offset,
                                                   int batch_size) {
-  kernel::sampling_from_logits_kernel<BLOCK_THREADS,
-                                      VEC_SIZE,
-                                      DType,
-                                      IdType>(
+  kernel::sampling_from_logits_kernel<BLOCK_THREADS, VEC_SIZE, DType, IdType>(
       logits, output, vocab_size, philox_seed, philox_offset, batch_size);
 }
